update transcription api key - refactor
