+++ src/renderer/components/ProjectCreation/ModalContainer.tsx
@@ -121,7 +121,6 @@ const ModalContainer = ({ isOpen, closeModal }: Props) => {
             nextView={nextView}
             projectName={projectName}
             textToDisplay={null}
-            open
           />
         );
       case ImportMediaView:
